Crate query_core

Re-exports

pub use error::*;

pub use executor::*;

pub use interpreter::*;

pub use query_ast::*;

pub use query_document::*;

pub use query_graph::*;

pub use query_graph_builder::*;

pub use response_ir::*;

pub use result_ast::*;

pub use schema::*;

pub use schema_builder::*;

Modules

error

executor

What the executor module DOES:

interpreter

query_ast

query_document

Intermediate representation of the input document that is used by the query engine to build query ASTs and validate the incoming data.

query_graph

query_graph_builder

Query graph builder module. tbd

response_ir

Prisma Response IR (Intermediate Representation).

result_ast

schema

schema_builder

Query schema builder. Root for query schema building.

Type Definitions

Result

Result type tying all sub-result type hierarchies of the core together.