Enum query_core::query_graph::QueryGraphDependency

pub enum QueryGraphDependency {
    ExecutionOrder,
    ParentResult(ParentResultFn),
    ParentProjection(ModelProjection, ParentProjectionFn),
    Then,
    Else,
}

Stored on the edges of the QueryGraph, a QueryGraphDependency contains information on how children are connected to their parents, expressing for example the need for additional information from the parent to be able to execute at runtime.

Variants

ExecutionOrder

Simple dependency indicating order of execution. Effectively an ordering and reachability tool for now.

ParentResult(ParentResultFn)

Performs a transformation on the child node based on the raw parent query result. Does not move the result to the closure, but provides a reference.

Tuple Fields of ParentResult

0: ParentResultFn

ParentProjection(ModelProjection, ParentProjectionFn)

More specialized version of ParentResult with more guarantees and side effects.

Performs a transformation on the child node based on the requested projection of the parent result (represented as a single merged ModelProjection). Assumes that the parent result can be converted into the requested projection, else a runtime error will occur. The ModelProjection is used to determine the set of values to extract from the parent result.

Important note: As opposed to ParentResult, this dependency guarantees that if the closure is called, the parent result was projected successfully. To achieve that, the query graph is post-processed in the finalize and reloads are injected at points where a projection is not fulfilled. See insert_reloads for more information.

Tuple Fields of ParentProjection

0: ModelProjection1: ParentProjectionFn

Then

Only valid in the context of a If control flow node.

Else

Only valid in the context of a If control flow node.