Module query_core::query_graph_builder

Query graph builder module. tbd

Modules

write

Structs

QueryGraphBuilder

RelationViolation

Enums

QueryGraphBuilderError

Traits

Builder

Temporary trait for the legacy read builder code.

Functions

aggregate

extract_filter

Extracts a regular filter potentially matching many records.

extract_nested_rel_aggr_selections

extract_query_args

Expects the caller to know that it is structurally guaranteed that query arguments can be extracted, e.g. that the query schema guarantees that required fields are present. Errors occur if conversions fail.

extract_unique_filter

Extracts a filter for a unique selector, i.e. a filter that selects exactly one record.

find_first

find_many

find_related

find_unique

Builds a read query from a parsed incoming read query field.

group_by

is_aggr_selection

Type Definitions

QueryGraphBuilderResult

Query graph builder sub-result type.