Function query_core::query_graph_builder::write::utils::insert_emulated_on_update_with_intermediary_node

pub fn insert_emulated_on_update_with_intermediary_node(
    graph: &mut QueryGraph, 
    connector_ctx: &ConnectorContext, 
    model_to_update: &ModelRef, 
    parent_node: &NodeRef, 
    child_node: &NodeRef
) -> QueryGraphBuilderResult<Option<NodeRef>>

Inserts emulated referential actions for onUpdate into the graph. All relations that refer to the model row(s) being deleted are checked for their desired emulation and inserted accordingly. Right now, supported modes are Restrict and SetNull (cascade will follow). Those checks fail at runtime and are inserted between parent_node and child_node.

This function is usually part of a delete (deleteOne or deleteMany). Expects parent_node to return one or more IDs (for records of model) to be checked.

Resulting graph (all emulations):

   ┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
           Parent       │
   │  (ids to update)
    ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ┘
              │
              ▼
   ┌────────────────────┐
┌ ─│     Join node      │─────────────────┬─────────────────────────────┬────────────────────────────────────────┐
   └────────────────────┘                 │                             │                                        │
│             │                           │                             │                                        │
              │                           │                             │                                        │
│             ▼                           ▼                             ▼                                        ▼
   ┌────────────────────┐      ┌────────────────────┐        ┌────────────────────┐                   ┌────────────────────┐
│  │Find Connected Model│      │Find Connected Model│        │Find Connected Model│                   │Find Connected Model│
   │    A (Restrict)    │      │    B (Restrict)    │     ┌──│    C (SetNull)     │                ┌──│    D (Cascade)     │
│  └────────────────────┘      └────────────────────┘     │  └────────────────────┘                │  └────────────────────┘
              │                           │               │             │                          │             │
│      Fail if│> 0                 Fail if│> 0            │             ▼                          │             │
              │                           │               │┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─               │             ▼
│             ▼                           ▼               │  ┌────────────────────┐ │              │┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
   ┌────────────────────┐      ┌────────────────────┐     ││ │  Insert onUpdate   │                │  ┌────────────────────┐ │
│  │       Empty        │      │       Empty        │     │  │ emulation subtree  │ │              ││ │  Insert onUpdate   │
   └────────────────────┘      └────────────────────┘     ││ │for relations using │                │  │ emulation subtree  │ │
│             │                           │               │  │the foreign key that│ │              ││ │ for all relations  │
              │                           │               ││ │    was updated.    │                │  │   pointing to D.   │ │
│             │                           │               │  └────────────────────┘ │              ││ └────────────────────┘
              │                           │               │└ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─               │ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ┘
│             │                           │               │             │                          │             │
              │                           │               │             │                          │             │
│             ▼                           │               │             ▼                          │             ▼
   ┌ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─                  │               │  ┌────────────────────┐                │  ┌────────────────────┐
└ ▶        Update       │◀────────────────┘               │  │ Update Cs (set FK  │                └─▶│     Update Cs      │
   └ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─                                  └─▶│       null)        │                   └────────────────────┘
              ▲                                              └────────────────────┘                              │
              │                                                         │                                        │
              └─────────────────────────────────────────────────────────┴────────────────────────────────────────┘