Module query_core::response_ir

Prisma Response IR (Intermediate Representation).

This module takes care of processing query execution results and transforming them into a different AST.

This IR is meant for general processing and storage. It can also be easily serialized.

Note: The code itself can be considered WIP. It is clear when reading the code that there are missing abstractions and a restructure might be necessary (good example is the default value handling sprinkled all over the place).

Structs

List

ResponseData

Enums

Item

An IR item that either expands to a subtype or leaf-record.

Type Aliases

ItemRef

Convenience type wrapper for Arc.

Map

A key -> value map to an IR item