Crate query_core

Re-exports

pub extern crate schema;

pub extern crate schema_builder;

pub use error::*;

pub use executor::*;

pub use interactive_transactions::*;

pub use interpreter::*;

pub use query_ast::*;

pub use query_document::*;

pub use query_graph::*;

pub use query_graph_builder::*;

pub use response_ir::*;

pub use result_ast::*;

pub use trace_helpers::*;

Modules

error

executor

What the executor module DOES:

interactive_transactions

interpreter

query_ast

query_document

Intermediate representation of the input document that is used by the query engine to build query ASTs and validate the incoming data.

query_graph

query_graph_builder

Query graph builder module. tbd

response_ir

Prisma Response IR (Intermediate Representation).

result_ast

trace_helpers

Type Definitions

Result

Result type tying all sub-result type hierarchies of the core together.